requirements.txt
asgiref==3.5.2
Django==4.0.5
django-environ==0.9.0
django-phonenumber-field==6.3.0
phonenumbers==8.12.50
psycopg2==2.9.3
sqlparse==0.4.2
